+++ addons/phantom_camera/scripts/managers/phantom_camera_manager.gd
@@ -19,9 +19,8 @@ var phantom_camera_3ds: Array: ## Note: To support disable_3d export templates f
 var _phantom_camera_3d_list: Array ## Note: To support disable_3d export templates for 2D projects, this is purposely not strongly typed.
 
 func _ready():
-	if Engine.has_singleton(PHANTOM_CAMERA_CONSTS.PCAM_MANAGER_NODE_NAME):
-		Engine.unregister_singleton(PHANTOM_CAMERA_CONSTS.PCAM_MANAGER_NODE_NAME)
-	Engine.register_singleton(PHANTOM_CAMERA_CONSTS.PCAM_MANAGER_NODE_NAME, self)
+	if not Engine.has_singleton(PHANTOM_CAMERA_CONSTS.PCAM_MANAGER_NODE_NAME):
+		Engine.register_singleton(PHANTOM_CAMERA_CONSTS.PCAM_MANAGER_NODE_NAME, self)
 
 func _enter_tree():
 	Engine.physics_jitter_fix = 0
